requirements.txt
Flask==3.1.1
Pillow==10.3.0
